Clarify 2 byte color encoding
From the example I believe this is what was intended? A literal interpretation of the original text could be interpreted as saying the color values literally are just 4 bits per channel, but in 8 bit color space, so the brightest you can generate in each channel is 0x0F, which doesn't match the example.