Minor rendering optimizations
Turns out the `Path()` constructor was taking half the time of rendering an image. Using a static Path removes all of that time.

The other optimizations are less significant. `Path.close()` is expensive, but is pointless since we always fill (this may not be a real savings, since the work presumably is just done when painting now, but it does get it off the profile). The `nextByte` getter is very hot; inlining `nextByte` seemed to help, but I might be imagining it. The real problem here is that what we really want to do is just walk a buffer with a pointer but Dart doesn't give us that option short of playing with dart:ffi and that seems unnecessarily complicated...